Split classes from core/events/variable_events.js into multiple files
