+++ core/events/events_var_create.js
@@ -1,0 +1,82 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @fileoverview Class for a variable creation event.
+ * @author fenichel@google.com (Rachel Fenichel)
+ */
+'use strict';
+
+goog.provide('Blockly.Events.VarCreate');
+
+goog.require('Blockly.Events');
+goog.require('Blockly.Events.VarBase');
+goog.require('Blockly.registry');
+goog.require('Blockly.utils.object');
+
+goog.requireType('Blockly.VariableModel');
+
+
+/**
+ * Class for a variable creation event.
+ * @param {!Blockly.VariableModel=} opt_variable The created variable. Undefined
+ *     for a blank event.
+ * @extends {Blockly.Events.VarBase}
+ * @constructor
+ */
+Blockly.Events.VarCreate = function(opt_variable) {
+  Blockly.Events.VarCreate.superClass_.constructor.call(this, opt_variable);
+  if (!opt_variable) {
+    return;  // Blank event to be populated by fromJson.
+  }
+
+  this.varType = opt_variable.type;
+  this.varName = opt_variable.name;
+};
+Blockly.utils.object.inherits(Blockly.Events.VarCreate, Blockly.Events.VarBase);
+
+/**
+ * Type of this event.
+ * @type {string}
+ */
+Blockly.Events.VarCreate.prototype.type = Blockly.Events.VAR_CREATE;
+
+/**
+ * Encode the event as JSON.
+ * @return {!Object} JSON representation.
+ */
+Blockly.Events.VarCreate.prototype.toJson = function() {
+  var json = Blockly.Events.VarCreate.superClass_.toJson.call(this);
+  json['varType'] = this.varType;
+  json['varName'] = this.varName;
+  return json;
+};
+
+/**
+ * Decode the JSON event.
+ * @param {!Object} json JSON representation.
+ */
+Blockly.Events.VarCreate.prototype.fromJson = function(json) {
+  Blockly.Events.VarCreate.superClass_.fromJson.call(this, json);
+  this.varType = json['varType'];
+  this.varName = json['varName'];
+};
+
+/**
+ * Run a variable creation event.
+ * @param {boolean} forward True if run forward, false if run backward (undo).
+ */
+Blockly.Events.VarCreate.prototype.run = function(forward) {
+  var workspace = this.getEventWorkspace_();
+  if (forward) {
+    workspace.createVariable(this.varName, this.varType, this.varId);
+  } else {
+    workspace.deleteVariableById(this.varId);
+  }
+};
+
+Blockly.registry.register(Blockly.registry.Type.EVENT,
+    Blockly.Events.VAR_CREATE, Blockly.Events.VarCreate);
